make functions mod move old string functions into their own submod make a hash submod with hash functions
